teller_mvp/api/requirements.txt
fastapi==0.110.0
uvicorn[standard]==0.29.0
httpx==0.27.0
python-dotenv==1.0.1
